requirements.txt
pydantic
langchain >= 0.1.13
langchain-community >= 0.0.20
langchain-core
langchain-groq
langchain-ollama
ipykernel 
python-dotenv 
elasticsearch
streamlit
pandas
pypdf
PyPDF2
chromadb
faiss-cpu
langchain_chroma
langchain_huggingface
sentence_transformers
grandalf
youtube-transcript-api
pinecone-client 
pinecone-text 
pinecone-notebooks
psycopg
psycopg2
pgvector 
faker 
google-generativeai
langchain-google-genai
mysql-connector-python